Add PC_SetPCId(pcpatch, int, bool, float8)
This commit adds a `PC_SetPCId` function that can be used to change the schema of a patch. The signature of the function is:

    PC_SetPCId(patch pcpatch, pcid int4, reinterpret boolean default false, defaultvalue float8 default 0.0)

The `patch` and `pcid` arguments are required, while `reinterpret` and `defaultvalue` are optional. When `TRUE` `reinterpret` indicates that the patch data may be re-interpreted, and that the data in the output patch may not be the same as the data in the input patch. `defaultvalue` is the value to used for dimensions that are in the new schema but not in the old schema. `defaultvalue` has not effect when `reinterpret` is `FALSE`.